fix(nvim): fix building neovim
A new build time dependency was adder for lua stuff:
luarocks_bootstrap. But it didn't build correctly with my overriden
luajit. The issue is this line in postInstall phase:

    substituteInPlace $out/etc/luarocks/* \
     --replace-fail '${lua.luaOnBuild}' '${lua}'

It is referring to luaOnBuild attribute from lua. It turns out that
luajit in that attribute has original src. Investigating further I found
out that the code that's supposed to override it along with the package
itself doesn't override the src attribute. To get passthru attrs a
passthruFun is called in this file:

    pkgs/development/interpreters/luajit/default.nix

But the override code is skipping all derivation attributes:

    inputs' = lib.filterAttrs (n: v: ! lib.isDerivation v && n != "passthruFun") inputs;
    override = attr: let lua = attr.override (inputs' // { self = lua; }); in lua;

src is a derivation too, so I conclude that's the reason why luaOnBuild
has original src. passthruFun is defined in

    pkgs/development/interpreters/lua-5/default.nix

To fix this issue, I overrode the src in pkgsBuildHost passed to luajit.