feat(nvim): enable format on save for bash
New version of bash-language-server has builtin support for shfmt. Good
thing about it is that it knows formatting options of the lsp client.
shfmt is using tabs by default, this was the reason why I didn't enable
it earlier (shfmt formatted with tabs if no editorconfig). With new
bash-language-server formatting support it formats with correct
indentation. And also supports editorconfig too: so additional
formatting options still can be configured with editorconfig.