fix(nvim): hide cursor in inactive terminal
The Catppuccin colorscheme defines the TermCursorNC highlight group, which
leaves a gray cursor in inactive terminal window. This can be
confusing. To resolve this, clear the TermCursorNC highlight
group to make the cursor invisible again. This behavior is similar to
tmux splits or terminal splits, where there is only one cursor on the
active pane.