refactor(flake): unpin nixpkgs
Previously, I pinned nixpkgs and updated it manually to the latest
commit used by nix channels. However, I discovered that I can simply use
the nixpkgs-unstable or nixos-unstable branches, as they are
automatically updated to the latest commits from the nix channels. I
opted for nixos-unstable instead of nixpkgs-unstable because it is what
is used in home-manager and nixvim flakes.