fix(flake): update home-manager URL
Since the commit where I began managing the Nix registry with
home-manager also on my dev PC, it has been pinned to the commit
specified in flake.lock. However, due to the use of an indirect URL,
`nix flake update` has stopped updating home-manager. This commit uses
the direct URL for home-manager to resolve the issue.